+++ kernel-rs/src/arena.rs
@@ -6,10 +6,10 @@ use core::pin::Pin;
 use array_macro::array;
 use pin_project::pin_project;
 
-use crate::list::*;
 use crate::lock::{Spinlock, SpinlockGuard};
-use crate::pinned_array::IterPinMut;
-use crate::rc_cell::{RcCell, Ref, RefMut};
+use crate::utils::list::*;
+use crate::utils::pinned_array::IterPinMut;
+use crate::utils::rc_cell::{RcCell, Ref, RefMut};
 
 /// A homogeneous memory allocator, equipped with the box type representing an allocation.
 pub trait Arena: Sized {
