+++ kernel-rs/src/lib.rs
@@ -67,22 +67,17 @@ mod file;
 mod fs;
 mod kalloc;
 mod kernel;
-mod list;
 mod lock;
 mod memlayout;
 mod page;
 mod param;
-mod pinned_array;
 mod pipe;
 mod plic;
 mod poweroff;
 mod proc;
-mod rc_cell;
 mod riscv;
 mod start;
 mod syscall;
-mod sysfile;
-mod sysproc;
 mod trap;
 mod uart;
 mod utils;
