+++ kernel-rs/src/syscall/mod.rs
@@ -9,6 +9,9 @@ use crate::{
     vm::{Addr, UVAddr},
 };
 
+mod file;
+mod proc;
+
 impl Kernel {
     pub fn syscall(&'static self, num: i32, proc: &mut CurrentProc<'_>) -> Result<usize, ()> {
         match num {
