+++ kernel-rs/src/utils/list.rs
