+++ kernel-rs/src/utils/rc_cell.rs
